app/requirements.txt
Django==1.9.3
Fabric==1.10.2
django-webpack-loader==0.2.4
djangorestframework==3.5.3
markdown=2.6.7
django-filter=1.0.1